+++ services/madoc-ts/src/frontend/shared/caputre-models/new/components/DefaultAdjacentNavigation.tsx
@@ -1,9 +1,19 @@
 import { CardButton, CardButtonGroup } from '@capture-models/editor';
 import React from 'react';
+import { useTranslation } from 'react-i18next';
+import styled from 'styled-components';
 import { useCurrentEntity } from '../hooks/use-current-entity';
 import { useFieldListNavigation } from '../hooks/use-field-list-navigation';
 
+const NoOverflowCardButton = styled(CardButton)`
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
+  box-shadow: none;
+`;
+
 export const DefaultAdjacentNavigation: React.FC = ({ children }) => {
+  const { t } = useTranslation();
   const [entity] = useCurrentEntity();
   const { goPrev, goNext } = useFieldListNavigation();
 
@@ -11,12 +21,12 @@ export const DefaultAdjacentNavigation: React.FC = ({ children }) => {
     <>
       {goNext || goPrev ? (
         <CardButtonGroup>
-          <CardButton size="small" onClick={goPrev} disabled={!goPrev}>
-            Previous {entity.label}
-          </CardButton>
-          <CardButton size="small" onClick={goNext} disabled={!goNext}>
-            Next {entity.label}
-          </CardButton>
+          <NoOverflowCardButton size="small" onClick={goPrev} disabled={!goPrev}>
+            {t('Prev {{label}}', { label: entity.label })}
+          </NoOverflowCardButton>
+          <NoOverflowCardButton size="small" onClick={goNext} disabled={!goNext}>
+            {t('Next {{label}}', { label: entity.label })}
+          </NoOverflowCardButton>
         </CardButtonGroup>
       ) : null}
       {children}